Cartesian coordinates with specialised grid for trees

This coord is based on the default Cartesian coordinates, but draws the a filled background in addition to the normal grid lines. The grid is forced to appear on every integer value within the scale's range.

Usage

coord_tree(
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  default = FALSE,
  clip = "on"
)

Arguments

xlim, ylim, expand, default, clip

See ggplot2::coord_cartesian()

Value

coord suitable for adding to a plot

Details

The appearance of the grid can be controlled with theme elements:

phylepic.grid.bar

filled grid (element_rect()).

phylepic.grid.line

grid line (element_line()).

phylepic.grid.every

grid frequency (integer). Default for both phylepic.grid.every.bar and phylepic.grid.every.stripe

phylepic.grid.every.bar

grid bar frequency (integer). Defaults to 2 to give an alternative striped background

phylepic.grid.every.stripe

grid bar frequency (integer). Defaults to 1 so that every tip on a tree has its own line