Plot calendar panel

Plot calendar panel

Usage

plot_calendar(
  phylepic,
  fill = NULL,
  weeks = FALSE,
  week_start = getOption("phylepic.week_start"),
  binned = TRUE,
  labels = NULL,
  labels.params = list(size = 3, fontface = "bold", colour = "white")
)

Arguments

phylepic

Object of class "phylepic".

fill

Variable in metadata table to use for the fill aesthetic (tidy-eval).

weeks

When TRUE, bin the date axis by weeks. Replaced by binned = TRUE paired with a suitable date scale.

week_start

See week_breaks().

binned

When TRUE, bin the date axis by the scale breaks.

labels

Controls the format of date labels on calendar tiles. If NULL, no labels are drawn. If a character scalar, controls the date format (see strptime()).

labels.params

Passed to ggplot2::geom_text() if labels are drawn.

Value

If phylepic is specified returns a ggplot, otherwise a function that when passed a "phylepic" object produces a ggplot for use with plot.phylepic().

See also